+++ pkg/key/app.go
@@ -86,6 +86,10 @@ func IsDeleted(customResource v1alpha1.App) bool {
 	return customResource.DeletionTimestamp != nil
 }
 
+func KubeConfigContextName(customResource v1alpha1.App) string {
+	return customResource.Spec.KubeConfig.Context.Name
+}
+
 func KubeConfigFinalizer(customResource v1alpha1.App) string {
 	return fmt.Sprintf("app-operator.giantswarm.io/app-%s", customResource.GetName())
 }