+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add key functions for cluster configmap and cluster kubeconfig names.
+
 ## [3.1.1] - 2020-11-10
 
 ### Fixed